[macos] mplayer-svn:build-macos.sh one step build
Refactored old script and README to reflect a new one-step mplayer 1.4 build script.
The script makes use of a necessary patch to define some openssl deprecated symbols
as it now builds with OpenSSL 1.1.1c and the latest ffmpeg 4.2 code.

Old Mplayer 1.1 cleanup.

In order to generate flags for enabled and disabled codecs programatically
a simple tool was created in go which outputs the bash code to set the flags
in bash variables which are evaluated by the build-macos.sh